Blow off the dust from this old project
I've updated it to 2021 Rust edition and fixed all the warnings

After updating the dependencies I tried to make it work but
sdl2 gave me problems that I was'n prepared to solve today.
I'm on Windows these days (sorry, not sorry) and figuring out
how to manage C dependencies here was not something trivial.
For the better! I decided to switch to native Rust libs and found
Pixels, that advertises itself as suitable for emulators.
Sounds perfect.

Most of the code for creating the window and handling keybord was
lifted from an expample that came with Pixels.