Make the opcode executor methods return the new address
There used to be a comment saying "each of the following functions
MUST move `ca` pointer" but the past me hasn't yet learned the lesson
that wrting what MUST to happen in a comment won't guarntee that it
indeed happens. And sure enough, while doing this small refacroring
I found that the execute_a_opcode didn't move the pointer.